Recent-Files: Track most recent files
When working with large remote repositories, its becomes hard to
navigate back to files that were previously opened and close (especially
if the repo has large tree structure with nested folders). Recent-Files
will locally track the set of open files. When a file is opened, it will
be added in Recent-Files. If the file window is closed, it can be
opened back from Recent-Files.

Recent-Files is an option that is disabled by default in the main config.